backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
playwright
groq
instagrapi
apscheduler
gspread
google-auth
python-telegram-bot
